pyproject.toml
[project]
name = "testfastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.129.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "pymorphy2>=0.9.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
]
